requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.5
django-bootstrap5==22.2
django-cleanup==6.0.0
django-environ==0.9.0
gunicorn==20.1.0
Pillow
pillow-heif
psycopg2
sqlparse==0.4.3
whitenoise==6.3.0
